### 2025-06-21 #8435 @crvernon
@ngreifer you have the license listed as GPL in your description file but we are going to need a copy of the actual license in the root of your repository.
