### 2025-06-21 #8435 @ngreifer
Just added the license file, thanks!
