+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   fwb-rep: fwb-rep.html
-last_built: 2025-06-16T15:29Z
+last_built: 2025-06-21T20:28Z
 urls:
   reference: https://ngreifer.github.io/fwb/reference
   article: https://ngreifer.github.io/fwb/articles
