### 2025-06-26 3dd45ea
ci level can now be as low as 0, added new wald interval, typo fixes
