### 2025-07-07 00b5a30
Version 0.5.0 updates: Simultaneous inference P-values as inversion of CIs Wald CIs fwb.array() tidy() method New tests
