+++ DESCRIPTION
@@ -38,7 +38,7 @@ Suggests:
     rmarkdown
 License: GPL (>= 2)
 Encoding: UTF-8
-URL: https://ngreifer.github.io/fwb/, https://github.com/ngreifer/fwb,
+URL: https://ngreifer.github.io/fwb/, https://github.com/ngreifer/fwb
 BugReports: https://github.com/ngreifer/fwb/issues
 RoxygenNote: 7.3.2
 Roxygen: list(markdown = TRUE)
