### 2025-07-08 aa94d6f
Linting and cleaning
