### 2025-08-11 302c9f6
Bug fix for findInterval (#6)
