### 2025-08-27 #8880 @editorialbot
[REVIEW]: fwb: An R package for the fractional weighted bootstrap
**Submitting author:** @ngreifer (<a href="http://orcid.org/0000-0003-3067-7154">Noah Greifer</a>)
**Repository:** https://github.com/ngreifer/fwb/
**Branch with paper.md** (empty if default branch):
**Version:** 0.7.0
