+++ R/set_fwb_wtype.R
@@ -11,7 +11,7 @@
 #'
 #' @seealso [fwb()] for a definition of each types of weights; [vcovFWB()]; [options()]; \pkgfun{boot}{boot} for the traditional bootstrap.
 #'
-#' @examplesIf requireNamespace("survival", quietly = TRUE)
+#' @examplesIf rlang::is_installed("survival")
 #' # Performing a Weibull analysis of the Bearing Cage
 #' # failure data as done in Xu et al. (2020)
 #' set.seed(123, "L'Ecuyer-CMRG")
