+++ man/fwb.Rd
@@ -114,7 +114,7 @@ Draws weights from a modification of the distribution described by Mammen (1983)
 
 }}
 \examples{
-\dontshow{if (requireNamespace("survival", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (rlang::is_installed("survival")) withAutoprint(\{ # examplesIf}
 # Performing a Weibull analysis of the Bearing Cage
 # failure data as done in Xu et al. (2020)
 set.seed(123, "L'Ecuyer-CMRG")
