+++ NEWS.md
@@ -1,9 +1,9 @@
 `cobalt` News and Updates
 ======
 
-# `fwb` (development version)
+# `fwb` 0.5.1
 
-* Fixed a bug where computing confidence intervals would yield an error about unused arguments for R version prior to 4.5.0. Thanks to @vincentarelbundock. (#6)
+* Fixed a bug where computing confidence intervals would yield an error about unused arguments for R versions prior to 4.5.0. Thanks to @vincentarelbundock. (#6)
 
 # `fwb` 0.5.0
 
