### 2025-11-04 #8880 @spholmes
Hi @prdm0:
Here is a bootstrap package in R that I really need a hand reviewing, any possibility you could help me out here?
Many thanks
Susan
