### 2025-11-09 #8880 @tobiasschoch
Dear @ngreifer,

I can install the package on both Windows 11 and Linux (openSUSE, Leap 15.6).

The installation results on CRAN produce a warning on `r-oldrel` versions (Mac and Windows); see https://cran.r-project.org/web/checks/check_results_fwb.html. The warning refers to the new arguments that were added to the function `findInterval()` in R version 4.5.0; see call of `findInterval()` in your function `norm_inter_invert()`, here: https://github.com/ngreifer/fwb/blob/master/R/compute.ci.R.

Best, Tobias
