### 2025-11-10 ac07ca5
Added new wtypes "beta" and "power"
