+++ reference/vcovFWB.md
@@ -1,0 +1,248 @@
+# Fractional Weighted Bootstrap Covariance Matrix Estimation
+
+`vcovFWB()` estimates the covariance matrix of model coefficient
+estimates using the fractional weighted bootstrap. It serves as a
+drop-in for [`stats::vcov()`](https://rdrr.io/r/stats/vcov.html) or
+[`sandwich::vcovBS()`](https://sandwich.R-Forge.R-project.org/reference/vcovBS.html).
+Clustered covariances are can be requested.
+
+## Usage
+
+``` r
+vcovFWB(
+  x,
+  cluster = NULL,
+  R = 1000,
+  start = FALSE,
+  wtype = getOption("fwb_wtype", "exp"),
+  ...,
+  fix = FALSE,
+  use = "pairwise.complete.obs",
+  .coef = stats::coef,
+  verbose = FALSE,
+  cl = NULL
+)
+```
+
+## Arguments
+
+- x:
+
+  a fitted model object, such as the output of a call to
+  [`lm()`](https://rdrr.io/r/stats/lm.html) or
+  [`glm()`](https://rdrr.io/r/stats/glm.html). The model object must
+  result from a function that can be updated using
+  [`update()`](https://rdrr.io/r/stats/update.html) and has a `weights`
+  argument to input non-integer case weights.
+
+- cluster:
+
+  a variable indicating the clustering of observations, a `list` (or
+  `data.frame`) thereof, or a formula specifying which variables from
+  the fitted model should be used (see examples). By default
+  (`cluster = NULL`), either `attr(x, "cluster")` is used (if any) or
+  otherwise every observation is assumed to be its own cluster.
+
+- R:
+
+  the number of bootstrap replications.
+
+- start:
+
+  `logical`; should `.coef(x)` be passed as `start` to the
+  `update(x, weights = ...)` call? In case the model `x` is computed by
+  some numeric iteration, this may speed up the bootstrapping.
+
+- wtype:
+
+  string; the type of weights to use. Allowable options include `"exp"`
+  (the default), `"pois"`, `"multinom"`, and `"mammen"`. See
+  [`fwb()`](https://ngreifer.github.io/fwb/reference/fwb.md) for
+  details. See
+  [`set_fwb_wtype()`](https://ngreifer.github.io/fwb/reference/set_fwb_wtype.md)
+  to set a global default.
+
+- ...:
+
+  ignored.
+
+- fix:
+
+  `logical`; if `TRUE`, the covariance matrix is fixed to be positive
+  semi-definite in case it is not.
+
+- use:
+
+  `character`; specification passed to
+  [`stats::cov()`](https://rdrr.io/r/stats/cor.html) for handling
+  missing coefficients/parameters.
+
+- .coef:
+
+  a function used to extract the coefficients from each fitted model.
+  Must return a numeric vector. By default,
+  [`stats::coef`](https://rdrr.io/r/stats/coef.html) is used, but
+  `marginaleffects::get_coef` can be a more reliable choice for some
+  models that have a non-standard
+  [`coef()`](https://rdrr.io/r/stats/coef.html) method, like that for
+  [`nnet::multinom()`](https://rdrr.io/pkg/nnet/man/multinom.html)
+  models.
+
+- verbose:
+
+  `logical`; whether to display a progress bar.
+
+- cl:
+
+  a cluster object created by
+  [`parallel::makeCluster()`](https://rdrr.io/r/parallel/makeCluster.html),
+  an integer to indicate the number of child-processes (integer values
+  are ignored on Windows) for parallel evaluations, or the string
[…]
