### 2025-11-10 de4e63c
drop0 can be NA
