+++ R/vcovFWB.R
@@ -1,13 +1,14 @@
 #' Fractional Weighted Bootstrap Covariance Matrix Estimation
 #'
-#' `vcovFWB()` estimates the covariance matrix of model coefficient estimates using the fractional weighted bootstrap. It serves as a drop-in for `stats::vcov()` or `sandwich::vcovBS()`. Clustered covariances are can be requested.
+#' `vcovFWB()` estimates the covariance matrix of model coefficient estimates using the fractional weighted bootstrap. It serves as a drop-in for `[stats::vcov()]` or `sandwich::vcovBS()`. Clustered covariances are can be requested.
 #'
 #' @inheritParams sandwich::vcovBS
 #' @inheritParams fwb
 #' @param x a fitted model object, such as the output of a call to `lm()` or `glm()`. The model object must result from a function that can be updated using [update()] and has a `weights` argument to input non-integer case weights.
-#' @param R the number of bootstrap replications.
-#' @param start `logical`; should `.coef(x)` be passed as `start` to the `update(x, weights = ...)` call? In case the model `x` is computed by some numeric iteration, this may speed up the bootstrapping.
-#' @param wtype string; the type of weights to use. Allowable options include `"exp"` (the default), `"pois"`, `"multinom"`, and `"mammen"`. See [fwb()] for details. See [set_fwb_wtype()] to set a global default.
+#' @param R the number of bootstrap replications. Default is 1000 (more is better but slower).
+#' @param start `logical`; should `.coef(x)` be passed as `start` to the `update(x, weights = ...)` call? In case the model `x` is computed by some numeric iteration, this may speed up the bootstrapping. Default is `FALSE`.
+#' @param wtype string; the type of weights to use. Allowable options include `"exp"` (the default), `"pois"`, `"multinom"`, `"mammen"`, `"beta"`, and `"power"`. See [fwb()] for details. See [set_fwb_wtype()] to set a global default.
+#' @param drop0 `logical`; when `wtype` is `"multinom"` or `"poisson"`, whether to drop units that are given weights of 0 from the model call in each iteration. If `TRUE`, the model will be called with an additional `subset` argument, filtering out units with weights of 0 (note this will overwrite any argument to `subset` in the original call). If `NA`, weights of 0 will be set to `NA` instead. Ignored for other `wtype`s because they don't produce 0 weights. Default is `FALSE`.
 #' @param ... ignored.
 #' @param fix `logical`; if `TRUE`, the covariance matrix is fixed to be positive semi-definite in case it is not.
 #' @param use `character`; specification passed to [stats::cov()] for handling missing coefficients/parameters.
@@ -15,13 +16,18 @@
 #'
 #' @inherit sandwich::vcovBS return
 #'
-#' @details `vcovFWB()` functions like other `vcov()`-like functions, such as those in the \pkg{sandwich} package, in particular, \pkgfun{sandwich}{vcovBS}, which implements the traditional bootstrap (and a few other bootstrap varieties for linear models). Sets of weights are generated as described in the documentation for [fwb()], and the supplied model is re-fit using those weights. When the fitted model already has weights, these are multiplied by the bootstrap weights.
+#' @details
[…]
