+++ vignettes/fwb-rep.Rmd
@@ -31,7 +31,7 @@ f.out <- fwb(.)
 
 ### Case 2: `simple = FALSE`, non-random `statistic`
 
-If `simple = FALSE` and `statistic` does not have a random component, see Case 1, regardless of whether or how parallelization is used. In this case, no random process occurs within each cluster, so no special steps need to be taken beyond setting a seed. Note that `simple` is `TRUE` by default unless `wtype = "multinom"`, so this must be set manually. See below for a code example:
+If `simple = FALSE` and `statistic` does not have a random component, see Case 1, regardless of whether or how parallelization is used. In this case, no random process occurs within each cluster, so no special steps need to be taken beyond setting a seed. Note that `simple = TRUE` by default unless `wtype = "multinom"`, so this must be set manually. See below for a code example:
 
 ```r
 set.seed({N})
@@ -84,7 +84,7 @@ There are some cases where you don't have to do any special work to ensure BCa i
 * `simple = TRUE`, there is no randomness in `statistic`, and no parallelization is used
 * `simple = TRUE`, there is no randomness in `statistic`, and `cl` is an integer or `"future"`
 
-In these cases, `fwb()` saves the state of the random seed that was used to originally generate the weights, and `fwb.ci()` recalls that seed to re-generate the weights and then computes the required statistics for the BCa interval without requiring any extra involvement by the user.
+In these cases, `fwb()` saves the state of the random seed that was used to originally generate the weights, and `fwb.ci()` recalls that seed to re-generate the weights and then computes the required statistics for the BCa interval without requiring any extra involvement by the user. Remember, `simple = TRUE` by default unless `wtype = "multinom"`.
 
 Otherwise, when the following condition is met, an additional step is required:
 
