### 2025-11-11 #8880 @ngreifer
Hi @tobiasschoch,

Thank you so much for the encouraging words about my documentation. It's something I care a lot about.

Regarding imported functions, I am following the advice in [*R Packages*](https://r-pkgs.org/) by Hadley Wickham and Jennifer Bryan. In their [section on importing functions](https://r-pkgs.org/dependencies-in-practice.html#sec-dependencies-in-imports-r-code), they recommend using `::` whenever possible and avoiding directly importing functions. I typically import all of the *stats* package because I consider it part of base R, but for other packages, I want to be explicit about where the function comes from. I would disagree that recording dependencies in `NAMESPACE` is an advantage; I know that when I read the source code of other packages, I prefer to know where a given function comes from rather than having to check the `NAMESPACE` to see if it was specifically imported (or even worse, if a whole package was imported and I am unable to tell where a given function comes from). As mentioned in the book, the efficiency cost of using `::` is on the order of 100ns.

Noah
