### 2025-11-16 #8880 @tobiasschoch
Dear @ngreifer,

From your answer, I learned that there are different opinions on that matter (usage of the `::` operator). I wasn't aware of Hadley's point of view.

Best, Tobias
