### 2025-12-04 #8880 @spholmes
Dear @scottkosty,
Would it be possible for you to give this submission to JOSS a shor treview, I have already secured one review but we like to have two, unfortunately at this time there are very few people available who know about the bootstrap and I know you care about open source.
Many thanks for the consideration,
Susan
