+++ man/fwb.Rd
@@ -111,7 +111,7 @@ Draws weights from a \eqn{\text{Beta}(\sqrt{2} - 1, 1)} distribution using \code
 }
 }
 
-\code{"exp"} is the default due to it being the formulation described in Xu et al. (2020) and in the most formulations of the Bayesian bootstrap; it should be used if one wants to remain in line with these guidelines or to maintain a Bayesian flavor to the analysis, whereas \code{"mammen"}, \code{"beta"}, and \code{"power"} might be preferred for their frequentist operating characteristics, though though more research is needed on their general performance. \code{"multinom"} and \code{"poisson"} should only be used for comparison purposes or as an alternative interface to \pkg{boot}.
+\code{"exp"} is the default due to it being the formulation described in Xu et al. (2020) and in the most formulations of the Bayesian bootstrap; it should be used if one wants to remain in line with these guidelines or to maintain a Bayesian flavor to the analysis, whereas \code{"mammen"}, \code{"beta"}, and \code{"power"} might be preferred for their frequentist operating characteristics, though more research is needed on their general performance. \code{"multinom"} and \code{"poisson"} should only be used for comparison purposes or as an alternative interface to \pkg{boot}.
 }
 }
 \section{Methods (by generic)}{
