+++ README.Rmd
@@ -14,35 +14,35 @@ knitr::opts_chunk$set(
 
 ```
 
-# `fwb`: Fractional Weighted Bootstrap
+# *fwb*: Fractional Weighted Bootstrap
 
 <!-- badges: start -->
 [![CRAN status](https://www.r-pkg.org/badges/version/fwb)](https://CRAN.R-project.org/package=fwb)
 [![CRAN_Downloads_Badge](https://cranlogs.r-pkg.org/badges/fwb)](https://CRAN.R-project.org/package=fwb)
 <!-- badges: end -->
 
-`fwb` implements the fractional weighted bootstrap (FWB), also known as the Bayesian bootstrap, following the treatment by Xu et al. (2020). The FWB involves generating sets of weights from a uniform Dirichlet distribution to be used in estimating statistics of interest, which yields a posterior distribution that can be interpreted in the same way the traditional (resampling-based) bootstrap distribution can be. The primary function is `fwb()`, which is essentially a drop-in for `boot::boot()` in that it takes in a dataset and a function and applies that function to the dataset and a randomly generated set of case weights. Also included are `fwb.ci()`, a drop-in for `boot::boot.ci()` for computing various kinds of confidence intervals (e.g., percentile, normal, bias-corrected percentile, etc.), and `vcovFWB()`, a drop-in for `sandwich::vcovBS()` for computing a coefficient covariance matrix from a regression model using the FWB.
+*fwb* implements the fractional weighted bootstrap (FWB), also known as the Bayesian bootstrap, following the treatment by Xu et al. (2020). The FWB involves generating sets of weights from a uniform Dirichlet distribution to be used in estimating statistics of interest, which yields a posterior distribution that can be interpreted in the same way the traditional (resampling-based) bootstrap distribution can be. The primary function is `fwb()`, which is essentially a drop-in for `boot::boot()` in that it takes in a dataset and a function and applies that function to the dataset and a randomly generated set of case weights. Also included are `fwb.ci()`, a drop-in for `boot::boot.ci()` for computing various kinds of confidence intervals (e.g., percentile, normal, bias-corrected percentile, etc.), and `vcovFWB()`, a drop-in for `sandwich::vcovBS()` for computing a coefficient covariance matrix from a regression model using the FWB.
 
-Check out the `fwb` [website](https://ngreifer.github.io/fwb/)!
+Check out the *fwb* [website](https://ngreifer.github.io/fwb/)!
 
 ## Installation
 
-You can install the current stable version of `fwb` from CRAN with:
+You can install the current stable version of *fwb* from CRAN with:
 
 ```{r, eval = FALSE}
 install.packages("fwb")
 ```
 
-You can install the development version of `fwb` from [GitHub](https://github.com/) with:
+You can install the development version of *fwb* from [GitHub](https://github.com/ngreifer/fwb) with:
 
 ```{r, eval = FALSE}
-# install.packages("remotes")
-remotes::install_github("ngreifer/fwb")
+# install.packages("pak")
+pak::pak("ngreifer/fwb")
 ```
 
 ## Examples
 
-Below are some examples of how to use `fwb`. We set a seed to ensure all results are replicable. (Note that when parallel processing is used, a special kind of seed needs to be set; see `vignette("fwb-rep")` for details.)
+Below are some examples of how to use *fwb*. We set a seed to ensure all results are replicable. (Note that when parallel processing is used, a special kind of seed needs to be set; see `vignette("fwb-rep")` for details.)
 
 ```{r}
 library("fwb")
@@ -137,24 +137,24 @@ It is clear that the estimates from the traditional bootstrap are pathological, 
 
 ## Weighted statistics and transformations
 
[…]
