+++ index.md
@@ -1,6 +1,6 @@
-# `fwb`: Fractional Weighted Bootstrap
+# *fwb*: Fractional Weighted Bootstrap
 
-`fwb` implements the fractional weighted bootstrap (FWB), also known as
+*fwb* implements the fractional weighted bootstrap (FWB), also known as
 the Bayesian bootstrap, following the treatment by Xu et al. (2020). The
 FWB involves generating sets of weights from a uniform Dirichlet
 distribution to be used in estimating statistics of interest, which
@@ -23,27 +23,27 @@ drop-in for
 for computing a coefficient covariance matrix from a regression model
 using the FWB.
 
-Check out the `fwb` [website](https://ngreifer.github.io/fwb/)!
+Check out the *fwb* [website](https://ngreifer.github.io/fwb/)!
 
 ## Installation
 
-You can install the current stable version of `fwb` from CRAN with:
+You can install the current stable version of *fwb* from CRAN with:
 
 ``` r
 install.packages("fwb")
 ```
 
-You can install the development version of `fwb` from
-[GitHub](https://github.com/) with:
+You can install the development version of *fwb* from
+[GitHub](https://github.com/ngreifer/fwb) with:
 
 ``` r
-# install.packages("remotes")
-remotes::install_github("ngreifer/fwb")
+# install.packages("pak")
+pak::pak("ngreifer/fwb")
 ```
 
 ## Examples
 
-Below are some examples of how to use `fwb`. We set a seed to ensure all
+Below are some examples of how to use *fwb*. We set a seed to ensure all
 results are replicable. (Note that when parallel processing is used, a
 special kind of seed needs to be set; see
 [`vignette("fwb-rep")`](https://ngreifer.github.io/fwb/articles/fwb-rep.md)
@@ -133,9 +133,9 @@ library("lmtest")
 # The traditional bootstrap fails
 coeftest(fit, vcov = sandwich::vcovBS)[1:3, ]
 #>              Estimate   Std. Error       z value Pr(>|z|)
-#> (Intercept) -6.904101 2.285991e+22 -3.020179e-22        1
-#> spontaneous  3.230286 1.670378e+14  1.933866e-14        1
-#> induced      2.190303 1.194912e+14  1.833025e-14        1
+#> (Intercept) -6.904101 2.191250e+14 -3.150760e-14        1
+#> spontaneous  3.230286 2.109943e+14  1.530982e-14        1
+#> induced      2.190303 1.382598e+14  1.584193e-14        1
 
 # The fractional weighted bootstrap succeeds
 coeftest(fit, vcov = vcovFWB)[1:3, ]
@@ -186,9 +186,9 @@ boot_est
 #> 
 #> Bootstrap Statistics :
 #>      original       bias     std. error
-#> t1* -6.904101 2.629510e+21 8.083197e+22
-#> t2*  3.230286 2.114656e+13 2.336509e+14
-#> t3*  2.190303 1.696351e+13 1.878001e+14
+#> t1* -6.904101 1.012278e+20 2.447447e+21
+#> t2*  3.230286 2.020323e+13 2.080529e+14
+#> t3*  2.190303 1.533283e+13 1.656799e+14
 
 fwb_est <- fwb(infert, fit_fun, R = 999, verbose = FALSE)
 fwb_est
@@ -228,7 +228,7 @@ bias-corrected percentile interval should probably be computed instead.
 
 ## Weighted statistics and transformations
 
-`fwb` also contains utility functions for computing weighted statistics
+*fwb* also contains utility functions for computing weighted statistics
 to facilitate incorporation of bootstrapped weights into estimates.
 These include
 [`w_mean()`](https://ngreifer.github.io/fwb/reference/w_mean.md),
[…]
