### 2026-01-20 #8880 @spholmes
> [@spholmes](https://github.com/spholmes) Thank you very much for thinking about me! This is an initiative I would love to contribute to. The next couple of months are busy for me but I could indeed do a short review by the end of January. Would that work? Best, Scott
>
> EDIT notes: I made an edit an hour after posting: I moved the time I could get a short review done to the end of January.

Hi @scottkosty : Just a reminder as this it has been a long time for the author to wait...and I really appreciate the help, if you could fill in the checklist items about COI that would ensure I don't have to find another reviewer, many thanks, Susan.
