### 2026-01-20 #8880 @spholmes
@scottkosty , you can generate your checklist with the command: `@editorialbot generate my checklist`...
