### 2026-01-27 8e1090a
Modified to use cli for errors
