+++ R/utils.R
@@ -45,7 +45,7 @@ check_index <- function(index, t, several.ok = FALSE) {
   if (ncol(t) == 1L) {
     if (!((is.numeric(index) && chk::vld_whole_number(index) && index == 1) ||
           (chk::vld_string(index) && !is_null(colnames(t)) && index == colnames(t)[1L]))) {
-      chk::wrn("only one statistic is available; ignoring `index`")
+      .wrn("only one statistic is available; ignoring {.arg index}")
     }
 
     return(1L)
@@ -53,100 +53,46 @@ check_index <- function(index, t, several.ok = FALSE) {
 
   if (!((is.character(index) || is.numeric(index)) && is_null(dim(index)))) {
     if (several.ok) {
-      .err("`index` must be a character or numeric vector indicating the names or indices of the desired statistics")
+      .err("{.arg index} must be a character or numeric vector indicating the names or indices of the desired statistics")
     }
     else {
-      .err("`index` must be a string or number indicating the name or index of the desired statistic")
+      .err("{.arg index} must be a string or number indicating the name or index of the desired statistic")
     }
   }
 
   index <- unique(drop(index))
 
   if (!several.ok && length(index) > 1L) {
-    .err("`index` must have length one")
+    .err("{.arg index} must have length one")
   }
 
   if (is.numeric(index)) {
     if (!chk::vld_whole_numeric(index)) {
       if (several.ok) {
-        .err("`index` must be a vector of positive integers")
+        .err("{.arg index} must be a vector of positive integers")
       }
       else {
-        .err("`index` must be a positive integer")
+        .err("{.arg index} must be a positive integer")
       }
     }
 
     if (any(index > ncol(t))) {
-      .err(sprintf("`index` must be between 1 and %s", ncol(t)))
+      .err(sprintf("{.arg index} must be between 1 and %s", ncol(t)))
     }
   }
   else {
     if (is_null(colnames(t))) {
-      .err("the estimates don't have names, so `index` must be numeric")
+      .err("the estimates don't have names, so {.arg index} must be numeric")
     }
 
     index <- match(index, colnames(t))
 
     if (anyNA(index)) {
-      .err("all entries in `index` must be the names of available statistics to compute.\n  The following are allowed: ",
-           toString(add_quotes(colnames(t), 2L)))
+      .err("all entries in {.arg index} must be the names of available statistics to compute. The following are allowed: {.val {colnames(t)}}")
     }
   }
 
   index
-}
-
-word_list <- function(word.list = NULL, and.or = "and", is.are = FALSE, quotes = FALSE) {
-  #When given a vector of strings, creates a string of the form "a and b"
-  #or "a, b, and c"
-  #If is.are, adds "is" or "are" appropriately
-
-  word.list <- setdiff(word.list, c(NA_character_, ""))
-
-  if (is_null(word.list)) {
-    out <- ""
-    attr(out, "plural") <- FALSE
-    return(out)
-  }
-
-  word.list <- add_quotes(word.list, quotes)
-
-  L <- length(word.list)
-
-  if (L == 1L) {
-    out <- word.list
-    if (is.are) out <- paste(out, "is")
-    attr(out, "plural") <- FALSE
-    return(out)
-  }
-
[…]
