### 2026-01-27 eb32900
Doc and metadata updates
