+++ man/macros/macros.Rd
@@ -1,1 +1,1 @@
-\newcommand{\pkgfun}{\ifelse{\Sexpr[results=rd,stage=render]{length(find.package("#1", quiet = TRUE, verbose = FALSE)) > 0}}{\code{\link[#1:#2]{#1::#2()}}}{\code{#1::#2()}}}
+\newcommand{\pkgfun}{\Sexpr[results=rd,stage=render]{if (rlang::is_installed("#1")) r"(\code{\link[#1:#2]{#1::#2()}})" else r"(\code{#1::#2()})"}}
