+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: fwb
 Type: Package
 Title: Fractional Weighted Bootstrap
-Version: 0.5.1.9001
+Version: 0.5.1.9002
 Authors@R: c(
     person("Noah", "Greifer", role=c("aut", "cre"), 
            email = "noah.greifer@gmail.com",
@@ -11,8 +11,9 @@ Description: An implementation of the fractional weighted bootstrap to be used a
 Depends:
     R (>= 4.0.0)
 Imports:
-    rlang (>= 1.1.6),
     chk (>= 0.10.0),
+    cli (>= 3.6.5),
+    rlang (>= 1.1.6),
     pbapply (>= 1.7-2),
     generics,
     graphics,
