+++ tests/testthat/helpers.R
@@ -36,8 +36,10 @@ expect_waldo_not_equal <- function(type, act, exp, info, ...) {
   invisible(act$val)
 }
 
-#Use regex to make strings invariant to white spaces
+#Use regex to make strings invariant to white spaces; also escape perens for perl
 .w <- function(x) {
+  x <- gsub("(", "\\(", x, fixed = TRUE)
+  x <- gsub(")", "\\)", x, fixed = TRUE)
   gsub(" ", "(\\s+)", x, fixed = TRUE)
 }
 
