+++ R/fwb.R
@@ -122,28 +122,28 @@ fwb <- function(data, statistic, R = 999, cluster = NULL, simple = NULL,
   bcall <- match.call()
 
   #Check data
-  chk::chk_not_missing(data, "`data`")
-  chk::chk_data(data)
+  arg::arg_supplied(data)
+  arg::arg_data.frame(data)
 
   n <- nrow(data)
-  if (is_null(n) || !chk::vld_count(n) || n < 1L) {
-    .err("{.arg data} must be present")
+  if (is_null(n) || !rlang::is_scalar_integerish(n) || n < 1L) {
+    arg::err("{.arg data} must be present")
   }
 
   #Check statistic
-  chk::chk_not_missing(statistic, "`statistic`")
-  chk::chk_function(statistic)
+  arg::arg_supplied(statistic)
+  arg::arg_function(statistic)
 
   #Check R
-  chk::chk_count(R)
-  chk::chk_gt(R, 0)
+  arg::arg_count(R)
+  arg::arg_gt(R, 0)
 
   #Check cluster
   clus <- substitute(cluster)
   cluster <- eval(clus, data, parent.frame())
 
   if (is_not_null(cluster)) {
-    .chk_atomic_vector(cluster)
+    arg::arg_vector(cluster)
 
     cluster <- factor(cluster)
     nc <- nlevels(cluster)
@@ -156,8 +156,8 @@ fwb <- function(data, statistic, R = 999, cluster = NULL, simple = NULL,
   strata_to_use <- NULL
 
   if (is_not_null(strata)) {
-    .chk_atomic_vector(strata)
-    chk::chk_length(strata, n)
+    arg::arg_vector(strata)
+    arg::arg_length(strata, n)
 
     strata_to_use <- factor(strata)
 
@@ -165,7 +165,7 @@ fwb <- function(data, statistic, R = 999, cluster = NULL, simple = NULL,
       cs <- unique(data.frame(cluster, strata_to_use))
 
       if (nrow(cs) != nlevels(cluster)) {
-        .err("clusters must be completely nested within strata")
+        arg::err("clusters must be completely nested within strata")
       }
 
       strata_to_use <- cs[[2L]]
@@ -177,7 +177,7 @@ fwb <- function(data, statistic, R = 999, cluster = NULL, simple = NULL,
     nw <- guess_num_workers(cl)
     verbose <- (nw == 1)
   }
-  chk::chk_flag(verbose)
+  arg::arg_flag(verbose)
 
   #Check wtype
   gen_weights <- make_gen_weights(wtype)
@@ -185,8 +185,9 @@ fwb <- function(data, statistic, R = 999, cluster = NULL, simple = NULL,
 
   #Check drop0
   if (wtype %in% c("multinom", "poisson")) {
-    chk::chk_scalar(drop0)
-    chk::chk_logical(drop0)
+    arg::arg_or(drop0,
+                arg::arg_is_NA,
+                arg::arg_flag)
   }
   else {
     drop0 <- FALSE
@@ -194,10 +195,10 @@ fwb <- function(data, statistic, R = 999, cluster = NULL, simple = NULL,
 
   #Check simple
   if (is_not_null(simple)) {
-    chk::chk_flag(simple)
+    arg::arg_flag(simple)
 
     # if (simple && wtype == "multinom") {
-    #   .err('`simple` cannot be `TRUE` when `wtype = "multinom"`')
+    #   arg::err('`simple` cannot be `TRUE` when `wtype = "multinom"`')
     # }
   }
   else {
@@ -211,7 +212,7 @@ fwb <- function(data, statistic, R = 999, cluster = NULL, simple = NULL,
 
[…]
