+++ DESCRIPTION
@@ -38,9 +38,9 @@ License: GPL (>= 2)
 Encoding: UTF-8
 URL: https://ngreifer.github.io/fwb/, https://github.com/ngreifer/fwb
 BugReports: https://github.com/ngreifer/fwb/issues
-RoxygenNote: 7.3.3
 Roxygen: list(markdown = TRUE)
 LazyData: true
 Config/testthat/edition: 3
 Config/testthat/parallel: false
 VignetteBuilder: knitr
+Config/roxygen2/version: 8.0.0

[…]
