+++ reference/fwb-package.md
@@ -29,3 +29,8 @@ Useful links:
 
 **Maintainer**: Noah Greifer <noah.greifer@gmail.com>
 ([ORCID](https://orcid.org/0000-0003-3067-7154))
+
+Authors:
+
+- Noah Greifer <noah.greifer@gmail.com>
+  ([ORCID](https://orcid.org/0000-0003-3067-7154))
