+++ reference/fwb.md
@@ -101,9 +101,8 @@ print(
   are ignored on Windows) for parallel evaluations, or the string
   `"future"` to use a `future` backend. See the `cl` argument of
   [`pbapply::pblapply()`](https://peter.solymos.org/pbapply/reference/pbapply.html)
-  for details. If `NULL`, no parallelization will take place. See
-  [`vignette("fwb-rep")`](https://ngreifer.github.io/fwb/articles/fwb-rep.md)
-  for details.
+  for details. If `NULL`, no parallelization will take place. See the
+  section "Parallel Processing" in Details.
 
 - ...:
 
@@ -125,7 +124,7 @@ print(
 
 ## Value
 
-An `fwb` object, which also inherits from `boot`, with the following
+An `<fwb>` object, which also inherits from `<boot>`, with the following
 components:
 
 - t0:
@@ -171,7 +170,7 @@ components:
 
   The type of weights used as determined by the `wtype` argument.
 
-`fwb` objects have [`coef()`](https://rdrr.io/r/stats/coef.html) and
+`<fwb>` objects have [`coef()`](https://rdrr.io/r/stats/coef.html) and
 [`vcov()`](https://rdrr.io/r/stats/vcov.html) methods, which extract the
 `t0` component and covariance of the `t` components, respectively.
 
@@ -308,6 +307,20 @@ provided close to nominal confidence interval coverage without
 excessively large intervals in the context of estimating the mean in a
 small sample. `"multinom"` and `"poisson"` should primarily be used for
 comparison purposes or as an alternative interface to boot.
+
+### Parallel Processing
+
+To speed up evaluation, parallel processing can be enabled. One way to
+do so is to supply an argument to `cl`. This can be either an
+integer(not available on Windows), a cluster object created by
+[`parallel::makeCluster()`](https://rdrr.io/r/parallel/makeCluster.html)
+, or the string `"future"`. Another general way is to use functionality
+in the [futurize](https://CRAN.R-project.org/package=futurize) package,
+which is compatible with fwb. See
+[`vignette("futurize-81-fwb", package = "futurize")`](https://futurize.futureverse.org/articles/futurize-81-fwb.html)
+for details. See also
+[`vignette("fwb-rep")`](https://ngreifer.github.io/fwb/articles/fwb-rep.md)
+for information on replicability with (and without) parallel processing.
 
 ## Methods (by generic)
 
