### 2026-08-12 b416f1d
Added cheap bootstrap CI from Lam (2022); minor code and doc updates
