### 2026-08-17 2824f1b
Recorded RNG streams, exhaustive multinomial bootstrap, test suite
Reproducibility

- fwb() and vcovFWB() reserve one L'Ecuyer-CMRG stream per replicate before any
  work is dispatched, so the weights depend only on the seed -- not on cl, the
  worker count, verbose, RNGkind(), or randomness in statistic. A plain
  set.seed() is enough everywhere, and clusterSetRNGStream() no longer has any
  effect. This changes estimates when simple = TRUE.
- fwb.array() recovers the weights without replaying the stream, so BCa
  intervals and weight recovery work in every case.
- <fwb> objects no longer sto
[…]
