+++ joss_paper/paper.md
@@ -17,7 +17,7 @@ bibliography: paper.bib
 
 # Summary
 
-The nonparametric bootstrap is a simple method for performing inference on estimates of statistical parameters without making strong assumptions about the data or a statistical model and without needing to derive the formula for an estimate's variance [@efronIntroductionBootstrap1993]. The traditional nonparametric bootstrap involves resampling the observations from the sample with replacement, computing the estimates in the bootstrap sample, performing these steps many times, and summarizing the resulting distribution of bootstrap estimates to compute their confidence intervals. This is equivalent to drawing integer weights from a multinomial distribution and computing weighted estimates in each bootstrap sample. However, integer weights are not the only weights that can be used, and fractional (non-integer) weights carry certain advantages, described below. `fwb` implements the fractional weighted bootstrap, also known as the Bayesian bootstrap, using a simple interface and with options for parallelization.
+The nonparametric bootstrap is a simple method for performing inference on estimates of statistical parameters without making strong assumptions about the data or a statistical model and without needing to derive the formula for an estimator's variance [@efronIntroductionBootstrap1993]. The traditional nonparametric bootstrap involves resampling the observations from the sample with replacement, computing the estimates in the bootstrap sample, performing these steps many times, and summarizing the resulting distribution of bootstrap estimates to compute their confidence intervals. This is equivalent to drawing integer weights from a multinomial distribution and computing weighted estimates in each bootstrap sample. However, integer weights are not the only weights that can be used, and fractional (non-integer) weights carry certain advantages, described below. `fwb` implements the fractional weighted bootstrap, also known as the Bayesian bootstrap, using a simple interface and with options for parallelization.
 
 # Statement of need
 
@@ -25,7 +25,7 @@ The fractional weighted bootstrap was described originally by @rubinBayesianBoot
 
 This is demonstrated in @xuApplicationsFractionalRandomWeightBootstrap2020a, who use the fractional weighted bootstrap to estimate the center and shape of a Weibull distribution characterizing failure times for a set of 1703 aircraft engines, among which only 6 failures were observed, data previously analyzed by @meekerStatisticalMethodsReliability1998. In cases with rare outcomes, rare covariate profiles, or highly influential observations, the traditional bootstrap can fail to yield valid inference. The fractional weighted bootstrap avoids these issues by retaining all units and allowing the weights to take on strictly positive, non-integer values.
 
[…]
