+++ joss_paper/paper.bib
@@ -8,7 +8,8 @@
   publisher={Chapman & Hall},
   author={Efron, Bradley and Tibshirani, Robert},
   year={1993},
-collection={Monographs on statistics and applied probability} }
+  doi={10.1201/9780429246593}
+}
 
 @article{rubinBayesianBootstrap1981,
 	title = {The Bayesian Bootstrap},
@@ -20,7 +21,6 @@ collection={Monographs on statistics and applied probability} }
 	volume = {9},
 	number = {1},
 	doi = {10.1214/aos/1176345338},
-	url = {https://www.jstor.org/stable/2240875},
 	note = {Publisher: Institute of Mathematical Statistics}
 }
 
@@ -35,9 +35,6 @@ collection={Monographs on statistics and applied probability} }
 	volume = {74},
 	number = {4},
 	doi = {10.1080/00031305.2020.1731599},
-	url = {https://doi.org/10.1080/00031305.2020.1731599},
-	note = {tex.ids= xuApplicationsFractionalRandomWeightBootstrap2020
-publisher: Taylor & Francis},
 	langid = {en}
 }
 
@@ -57,8 +54,7 @@ publisher: Taylor & Francis},
 	author = {Canty, Angelo and Ripley, Brian},
 	year = {2024},
 	publisher = {The R Foundation},
-	doi = {10.32614/cran.package.boot},
-	url = {http://dx.doi.org/10.32614/CRAN.package.boot}
+	doi = {10.32614/cran.package.boot}
 }
 
 @book{
@@ -71,15 +67,15 @@ publisher: Taylor & Francis},
   publisher={Cambridge University Press},
   author={Davison, Anthony Christopher and Hinkley, David Victor},
   year={1997},
-collection={Cambridge series on statistical and probabilistic mathematics} }
+  collection={Cambridge series on statistical and probabilistic mathematics}
+}
 
 @misc{solymos2023,
 	title = {pbapply: Adding Progress Bar to '*apply' Functions},
 	author = {Solymos, Peter and Zawadzki, Zygmunt},
 	year = {2023},
 	publisher = {The R Foundation},
-	doi = {10.32614/cran.package.pbapply},
-	url = {http://dx.doi.org/10.32614/CRAN.package.pbapply}
+	doi = {10.32614/cran.package.pbapply}
 }
 
 @article{bengtsson2021,
@@ -92,7 +88,6 @@ collection={Cambridge series on statistical and probabilistic mathematics} }
 	volume = {13},
 	number = {2},
 	doi = {10.32614/rj-2021-048},
-	url = {http://dx.doi.org/10.32614/RJ-2021-048},
 	langid = {en}
 }
 
@@ -105,8 +100,7 @@ collection={Cambridge series on statistical and probabilistic mathematics} }
 	journal = {The Annals of Statistics},
 	volume = {21},
 	number = {1},
-	doi = {10.1214/aos/1176349025},
-	url = {https://projecteuclid.org/journals/annals-of-statistics/volume-21/issue-1/Bootstrap-and-Wild-Bootstrap-for-High-Dimensional-Linear-Models/10.1214/aos/1176349025.full}
+	doi = {10.1214/aos/1176349025}
 }
 
 @article{hanley2006,
@@ -120,7 +114,6 @@ collection={Cambridge series on statistical and probabilistic mathematics} }
 	volume = {83},
 	number = {1},
 	doi = {10.1016/j.cmpb.2006.04.006},
-	url = {http://dx.doi.org/10.1016/j.cmpb.2006.04.006},
 	langid = {en}
 }
 
[…]
