+++ tests/testthat/test-fwb-ci.R
@@ -1,0 +1,555 @@
+#Confidence intervals: `fwb.ci()`, `get_ci()`, and the interval arithmetic in
+#`compute_ci()`.
+#
+#`test-boot.R` already checks agreement with `boot::boot.ci()` for the four interval
+#types the two packages share. What is left uncovered, and is what this file adds, is
+#the *fwb*-only machinery: the `"wald"`, `"bc"`, and `"cheap"` intervals, the
+#identities that tie the types together, and the argument handling.
+
+boot_for_ci <- function(n = 40L, R = 200L, seed = 81L) {
+  d <- fwb_test_df(n)
+
+  set.seed(seed, "L'Ecuyer-CMRG")
+  fwb(d, lm_stat, R = R, verbose = FALSE, simple = FALSE)
+}
+
+test_that("`fwb.ci()` returns one component per requested type", {
+  f <- boot_for_ci()
+
+  types <- c("perc", "bc", "wald", "norm", "cheap", "basic", "bca")
+
+  ci <- fwb.ci(f, type = types, index = 2L)
+
+  expect_s3_class(ci, c("fwbci", "bootci"))
+  expect_true(all(types %in% names(ci)))
+  expect_identical(ci[["R"]], f[["R"]])
+  expect_equal(ci[["t0"]], f[["t0"]][[2L]], tolerance = fwb_eps(),
+               ignore_attr = TRUE)
+  expect_identical(attr(ci, "conf", TRUE), .95)
+
+  #`"wald"`, `"norm"`, and `"cheap"` report level and the two limits; the
+  #order-statistic types additionally report the two ranks used.
+  for (ty in c("wald", "norm", "cheap")) {
+    expect_identical(dim(ci[[ty]]), c(1L, 3L), info = ty)
+  }
+
+  for (ty in c("perc", "bc", "basic", "bca")) {
+    expect_identical(dim(ci[[ty]]), c(1L, 5L), info = ty)
+  }
+
+  #`type = "all"` is the full set when there are no clusters.
+  ci_all <- fwb.ci(f, type = "all", index = 2L)
+
+  expect_true(all(types %in% names(ci_all)))
+})
+
+test_that("`type = \"all\"` drops BCa when clusters are present", {
+  d <- fwb_test_df()
+
+  set.seed(82, "L'Ecuyer-CMRG")
+  f <- fwb(d, lm_stat, R = 200L, verbose = FALSE, simple = FALSE, cluster = g)
+
+  ci <- fwb.ci(f, type = "all", index = 2L)
+
+  expect_false("bca" %in% names(ci))
+  expect_true(all(c("perc", "bc", "wald", "norm", "cheap", "basic") %in% names(ci)))
+})
+
+test_that("the interval types satisfy their defining identities", {
+  f <- boot_for_ci()
+
+  index <- 2L
+  t <- f[["t"]][, index]
+  t0 <- f[["t0"]][[index]]
+  conf <- .95
+  z <- qnorm(1 - (1 - conf) / 2)
+
+  ci <- fwb.ci(f, conf = conf, type = c("wald", "norm", "basic", "perc", "bc"),
+               index = index)
+
+  #Wald: t0 +/- z * sd, with no bias correction at all.
+  expect_equal(as.vector(ci[["wald"]][1L, 2:3]),
+               t0 + c(-1, 1) * z * sd(t),
+               tolerance = fwb_eps())
+
+  #Normal: the same, after subtracting the bootstrap bias.
+  expect_equal(as.vector(ci[["norm"]][1L, 2:3]),
+               t0 - (mean(t) - t0) + c(-1, 1) * z * sd(t),
+               tolerance = fwb_eps())
+
+  #Basic is the percentile interval reflected about t0, so the two must sum to 2*t0
+  #limit by limit (in reverse order).
[…]

[…]
