+++ news/index.md
@@ -2,6 +2,8 @@
 
 ## *fwb* (development version)
 
+#### New Features
+
 - Added a new confidence interval type for
   [`confint()`](https://rdrr.io/r/stats/confint.html),
   [`fwb.ci()`](https://ngreifer.github.io/fwb/reference/fwb.ci.md), and
@@ -11,9 +13,244 @@
   nominal coverage even with few bootstrap replications (as low as
   one!).
 
+#### Breaking changes
+
+- **Bootstrap weights now depend only on the seed.** Previously, when
+  `simple = TRUE`, the weights were drawn inside each bootstrap
+  replication and so depended on how the work was distributed: results
+  changed with `cl`, with the number of workers, and with `verbose`, and
+  reproducing a parallel run required `set.seed(###, "L'Ecuyer-CMRG")`
+  (or, with a `cluster` object,
+  [`parallel::clusterSetRNGStream()`](https://rdrr.io/r/parallel/RngStream.html)
+  instead of [`set.seed()`](https://rdrr.io/r/base/Random.html)).
+  [`fwb()`](https://ngreifer.github.io/fwb/reference/fwb.md) and
+  [`vcovFWB()`](https://ngreifer.github.io/fwb/reference/vcovFWB.md) now
+  reserve one random number stream per replicate before any work is
+  dispatched, and each replication draws its weights from its own
+  stream. A plain call to
+  [`set.seed()`](https://rdrr.io/r/base/Random.html) is all that is
+  needed, whatever `cl`, `verbose`, or
+  [`RNGkind()`](https://rdrr.io/r/base/Random.html) are set to, and
+  [`parallel::clusterSetRNGStream()`](https://rdrr.io/r/parallel/RngStream.html)
+  no longer has any effect. See
+  [`vignette("fwb-rep")`](https://ngreifer.github.io/fwb/articles/fwb-rep.md).
+
+- **This changes the bootstrap estimates when `simple = TRUE`** (the
+  default for every `wtype` except `"multinom"`). Analyses that need to
+  reproduce results from *fwb* 0.6.0 or earlier should install that
+  version (e.g., using `pak::pak("fwb@0.6.0")`). Results with
+  `simple = FALSE` are unchanged, so `wtype = "multinom"` continues to
+  match `boot::boot(., stype = "f")` exactly. Note also that
+  `simple = TRUE` and `simple = FALSE` no longer give the same estimates
+  as each other, even for the weight types where they previously did.
+
+- `<fwb>` objects no longer store the `cl` argument, since nothing about
+  the backend is needed to recover the weights any more. Previously, a
+  `cluster` object kept in the result made the object unusable after
+  [`parallel::stopCluster()`](https://rdrr.io/r/parallel/makeCluster.html)
+  or after being saved and reloaded.
+
+- [`vignette("fwb-rep")`](https://ngreifer.github.io/fwb/articles/fwb-rep.md)
+  has been rewritten around the new behavior. It no longer describes the
+  several cases that used to require different treatment, and it
+  explains how to reproduce results from earlier versions.
+
+- **The multinomial bootstrap is now exhaustive when it can be.** There
+  are only \\n^n\\ distinct multinomial bootstrap samples of \\n\\
+  units, so when `R` is at least that many,
+  [`fwb()`](https://ngreifer.github.io/fwb/reference/fwb.md) and
+  [`vcovFWB()`](https://ngreifer.github.io/fwb/reference/vcovFWB.md) now
+  use each of them exactly once instead of sampling, and set `R` to that
+  number (reporting this in a message). The estimates are then the exact
+  bootstrap distribution rather than a sample from it, and they no
[…]
