+++ reference/fwb.md
@@ -35,7 +35,8 @@ print(
 
 - data:
 
-  the dataset used to compute the statistic
+  the dataset used to compute the statistic; must contain more than one
+  unit
 
 - statistic:
 
@@ -43,19 +44,26 @@ print(
   the statistic(s) of interest. The function should take at least two
   arguments; the first argument should correspond to the dataset and the
   second argument should correspond to a vector of weights. Any further
-  arguments can be passed to `statistic` through the `...` argument.
+  arguments can be passed to `statistic` through the `...` argument, but
+  they cannot share a name with an argument of `fwb()`, which would
+  claim them instead. These requirements are checked unless `statistic`
+  has a `...` argument, in which case no assumption is made about what
+  it accepts.
 
 - R:
 
   the number of bootstrap replicates. Default is 999 but more is always
   better. For the percentile bootstrap confidence interval to be exact,
-  it can be beneficial to use one less than a multiple of 100.
+  it can be beneficial to use one less than a multiple of 100. When
+  `wtype = "multinom"` and `R` is at least the number of distinct
+  bootstrap samples of `data`, it is ignored; see the description of
+  `"multinom"` below.
 
 - cluster:
 
   optional; a vector containing cluster membership. If supplied, will
-  run the cluster bootstrap. See Details. Evaluated first in `data` and
-  then in the global environment.
+  run the cluster bootstrap. Must contain more than one cluster. See
+  Details. Evaluated first in `data` and then in the global environment.
 
 - simple:
 
@@ -103,8 +111,11 @@ print(
   are ignored on Windows) for parallel evaluations, or the string
   `"future"` to use a `future` backend. See the `cl` argument of
   [`pbapply::pblapply()`](https://peter.solymos.org/pbapply/reference/pbapply.html)
-  for details. If `NULL`, no parallelization will take place. See the
-  section "Parallel Processing" in Details.
+  for details. If `NULL`, no parallelization will take place.
+  Alternatively, the
+  [futurize](https://CRAN.R-project.org/package=futurize) package can be
+  used to incorporate a `future` backend. See the section "Parallel
+  Processing" in Details.
 
 - ...:
 
@@ -141,16 +152,13 @@ components:
 
 - R:
 
-  The value of `R` as passed to `fwb()`.
+  The value of `R` as passed to `fwb()`, except in the exhaustive
+  multinomial case described under `"multinom"` below, where it is the
+  number of distinct bootstrap samples.
 
 - data:
 
   The `data` as passed to `fwb()`.
-
-- seed:
-
-  The value of `.Random.seed` just prior to generating the weights
-  (after the first call to `statistic` with uniform weights).
 
 - statistic:
 
@@ -171,6 +179,12 @@ components:
 - wtype:
 
   The type of weights used as determined by the `wtype` argument.
+
+The seed (when `simple = FALSE`) or seeds (when `simple = TRUE`) used to
+re-generate the weights is stored in the `"seeds"` attribute of the
+returned object. In the exhaustive multinomial case described under
+`"multinom"` below, no random numbers are used and the `"exhaustive"`
+attribute is `TRUE`.
 
 `<fwb>` objects have [`coef()`](https://rdrr.io/r/stats/coef.html) and
 [`vcov()`](https://rdrr.io/r/stats/vcov.html) methods, which extract the
@@ -203,10 +217,10 @@ the `family` to a "quasi" variety (e.g.,
 [`quasibinomial()`](https://rdrr.io/r/stats/family.html)) to avoid a
 spurious warning about "non-integer \#successes".
 
-The cluster bootstrap can be requested by supplying a vector of cluster
[…]
