### 2026-08-21 72bec01
Fixed issues with empty random seed
