+++ README.Rmd
@@ -14,11 +14,10 @@ knitr::opts_chunk$set(
 
 ```
 
-# *fwb*: Fractional Weighted Bootstrap
+# *fwb*: Fractional Weighted Bootstrap <img src="man/figures/logo.png" align="right" width="150"/>
 
 <!-- badges: start -->
-[![CRAN status](https://www.r-pkg.org/badges/version/fwb)](https://CRAN.R-project.org/package=fwb)
-[![CRAN_Downloads_Badge](https://cranlogs.r-pkg.org/badges/fwb)](https://CRAN.R-project.org/package=fwb)
+[![CRAN_Status_Badge](https://r-pkg.org/badges/version/fwb?color=C6551E)](https://cran.r-project.org/package=fwb) [![CRAN_Downloads_Badge](https://cranlogs.r-pkg.org/badges/fwb?color=C6551E)](https://cran.r-project.org/package=fwb)
 <!-- badges: end -->
 
 *fwb* implements the fractional weighted bootstrap (FWB), also known as the Bayesian bootstrap, following the treatment by Xu et al. (2020). The FWB involves generating sets of weights from a continuous distribution to be used in estimating statistics of interest, which yields a posterior distribution that can be interpreted in the same way the traditional (resampling-based) bootstrap distribution can be. The primary function is `fwb()`, which is essentially a drop-in for `boot::boot()` in that it takes in a dataset and a function and applies that function to the dataset and a randomly generated set of case weights. Also included are `fwb.ci()`, a drop-in for `boot::boot.ci()` for computing various kinds of confidence intervals (e.g., percentile, normal, bias-corrected percentile, etc.), and `vcovFWB()`, a drop-in for `sandwich::vcovBS()` for computing a coefficient covariance matrix from a regression model using the FWB.
@@ -70,7 +69,7 @@ weibull_est <- function(data, w) {
 
 # 1999 bootstrap replications; more is always better
 fwb_est <- fwb(bearingcage, statistic = weibull_est,
-               R = 1999, verbose = FALSE)
+               R = 1999)
 fwb_est
 
 # Bias-corrected accelerated percentile
@@ -119,7 +118,7 @@ fit_fun <- function(data, w) {
 boot_est <- boot::boot(infert, fit_fun, R = 999, stype = "f")
 boot_est
 
-fwb_est <- fwb(infert, fit_fun, R = 999, verbose = FALSE)
+fwb_est <- fwb(infert, fit_fun, R = 999)
 fwb_est
 ```
 
