+++ man/fwb-package.Rd
@@ -5,6 +5,8 @@
 \alias{fwb-package}
 \title{fwb: Fractional Weighted Bootstrap}
 \description{
+\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
+
 An implementation of the fractional weighted bootstrap to be used as a drop-in for functions in the 'boot' package. The fractional weighted bootstrap (also known as the Bayesian bootstrap) involves drawing weights randomly that are applied to the data rather than resampling units from the data. See Xu et al. (2020) \doi{10.1080/00031305.2020.1731599} for details.
 }
 \references{
