+++ NEWS.md
@@ -1,7 +1,7 @@
 *fwb* News and Updates
 ======
 
-# *fwb* (development version)
+# *fwb* 0.7.0
 
 ### New Features
 
