+++ README.md
@@ -1,13 +1,12 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# *fwb*: Fractional Weighted Bootstrap
+# *fwb*: Fractional Weighted Bootstrap <img src="man/figures/logo.png" align="right" width="150"/>
 
 <!-- badges: start -->
 
-[![CRAN
-status](https://www.r-pkg.org/badges/version/fwb)](https://CRAN.R-project.org/package=fwb)
-[![CRAN_Downloads_Badge](https://cranlogs.r-pkg.org/badges/fwb)](https://CRAN.R-project.org/package=fwb)
+[![CRAN_Status_Badge](https://r-pkg.org/badges/version/fwb?color=C6551E)](https://cran.r-project.org/package=fwb)
+[![CRAN_Downloads_Badge](https://cranlogs.r-pkg.org/badges/fwb?color=C6551E)](https://cran.r-project.org/package=fwb)
 <!-- badges: end -->
 
 *fwb* implements the fractional weighted bootstrap (FWB), also known as
@@ -81,12 +80,12 @@ weibull_est <- function(data, w) {
 
 # 1999 bootstrap replications; more is always better
 fwb_est <- fwb(bearingcage, statistic = weibull_est,
-               R = 1999, verbose = FALSE)
+               R = 1999)
 fwb_est
 #> FRACTIONAL WEIGHTED BOOTSTRAP
 #> 
 #> Call:
-#> fwb(data = bearingcage, statistic = weibull_est, R = 1999, verbose = FALSE)
+#> fwb(data = bearingcage, statistic = weibull_est, R = 1999)
 #> 
 #> Bootstrap Statistics :
 #>          original         bias   std. error
@@ -185,12 +184,12 @@ boot_est
 #> t2*  3.230286  4.061618e+13 3.237388e+14
 #> t3*  2.190303  2.903494e+13 2.359882e+14
 
-fwb_est <- fwb(infert, fit_fun, R = 999, verbose = FALSE)
+fwb_est <- fwb(infert, fit_fun, R = 999)
 fwb_est
 #> FRACTIONAL WEIGHTED BOOTSTRAP
 #> 
 #> Call:
-#> fwb(data = infert, statistic = fit_fun, R = 999, verbose = FALSE)
+#> fwb(data = infert, statistic = fit_fun, R = 999)
 #> 
 #> Bootstrap Statistics :
 #>              original       bias std. error
