+++ reference/index.md
@@ -4,21 +4,29 @@
   [`print(`*`<fwb>`*`)`](https://ngreifer.github.io/fwb/reference/fwb.md)
   : Fractional Weighted Bootstrap
 
-- [`fwb.ci()`](https://ngreifer.github.io/fwb/reference/fwb.ci.md)
-  [`print(`*`<fwbci>`*`)`](https://ngreifer.github.io/fwb/reference/fwb.ci.md)
-  : Fractional Weighted Bootstrap Confidence Intervals
-
 - [`summary(`*`<fwb>`*`)`](https://ngreifer.github.io/fwb/reference/summary.fwb.md)
   [`confint(`*`<fwb>`*`)`](https://ngreifer.github.io/fwb/reference/summary.fwb.md)
   :
 
   Summarize `fwb` Output
 
+- [`vcovFWB()`](https://ngreifer.github.io/fwb/reference/vcovFWB.md) :
+  Fractional Weighted Bootstrap Covariance Matrix Estimation
+
+- [`fwb.ci()`](https://ngreifer.github.io/fwb/reference/fwb.ci.md)
+  [`print(`*`<fwbci>`*`)`](https://ngreifer.github.io/fwb/reference/fwb.ci.md)
+  : Fractional Weighted Bootstrap Confidence Intervals
+
 - [`plot(`*`<fwb>`*`)`](https://ngreifer.github.io/fwb/reference/plot.fwb.md)
   : Plots of the Output of a Fractional Weighted Bootstrap
 
-- [`vcovFWB()`](https://ngreifer.github.io/fwb/reference/vcovFWB.md) :
-  Fractional Weighted Bootstrap Covariance Matrix Estimation
+- [`get_ci()`](https://ngreifer.github.io/fwb/reference/get_ci.md) :
+
+  Extract Confidence Intervals from a `bootci` Object
+
+- [`set_fwb_wtype()`](https://ngreifer.github.io/fwb/reference/set_fwb_wtype.md)
+  [`get_fwb_wtype()`](https://ngreifer.github.io/fwb/reference/set_fwb_wtype.md)
+  : Set weights type
 
 - [`w_mean()`](https://ngreifer.github.io/fwb/reference/w_mean.md)
   [`w_var()`](https://ngreifer.github.io/fwb/reference/w_mean.md)
@@ -31,14 +39,6 @@
   [`w_scale()`](https://ngreifer.github.io/fwb/reference/w_mean.md)
   [`w_center()`](https://ngreifer.github.io/fwb/reference/w_mean.md) :
   Calculate weighted statistics
-
-- [`set_fwb_wtype()`](https://ngreifer.github.io/fwb/reference/set_fwb_wtype.md)
-  [`get_fwb_wtype()`](https://ngreifer.github.io/fwb/reference/set_fwb_wtype.md)
-  : Set weights type
-
-- [`get_ci()`](https://ngreifer.github.io/fwb/reference/get_ci.md) :
-
-  Extract Confidence Intervals from a `bootci` Object
 
 - [`fwb.array()`](https://ngreifer.github.io/fwb/reference/fwb.array.md)
   : Recover Bootstrap Weights
