+++ llms.txt
@@ -86,12 +86,12 @@ weibull_est <- function(data, w) {
 
 # 1999 bootstrap replications; more is always better
 fwb_est <- fwb(bearingcage, statistic = weibull_est,
-               R = 1999, verbose = FALSE)
+               R = 1999)
 fwb_est
 #> FRACTIONAL WEIGHTED BOOTSTRAP
 #> 
 #> Call:
-#> fwb(data = bearingcage, statistic = weibull_est, R = 1999, verbose = FALSE)
+#> fwb(data = bearingcage, statistic = weibull_est, R = 1999)
 #> 
 #> Bootstrap Statistics :
 #>          original         bias   std. error
@@ -197,12 +197,12 @@ boot_est
 #> t2*  3.230286  4.061618e+13 3.237388e+14
 #> t3*  2.190303  2.903494e+13 2.359882e+14
 
-fwb_est <- fwb(infert, fit_fun, R = 999, verbose = FALSE)
+fwb_est <- fwb(infert, fit_fun, R = 999)
 fwb_est
 #> FRACTIONAL WEIGHTED BOOTSTRAP
 #> 
 #> Call:
-#> fwb(data = infert, statistic = fit_fun, R = 999, verbose = FALSE)
+#> fwb(data = infert, statistic = fit_fun, R = 999)
 #> 
 #> Bootstrap Statistics :
 #>              original       bias std. error
@@ -327,21 +327,29 @@ Statistician*, 74(4), 345–358.
   [`print(`*`<fwb>`*`)`](https://ngreifer.github.io/fwb/reference/fwb.md)
   : Fractional Weighted Bootstrap
 
-- [`fwb.ci()`](https://ngreifer.github.io/fwb/reference/fwb.ci.md)
-  [`print(`*`<fwbci>`*`)`](https://ngreifer.github.io/fwb/reference/fwb.ci.md)
-  : Fractional Weighted Bootstrap Confidence Intervals
-
 - [`summary(`*`<fwb>`*`)`](https://ngreifer.github.io/fwb/reference/summary.fwb.md)
   [`confint(`*`<fwb>`*`)`](https://ngreifer.github.io/fwb/reference/summary.fwb.md)
   :
 
   Summarize `fwb` Output
 
+- [`vcovFWB()`](https://ngreifer.github.io/fwb/reference/vcovFWB.md) :
+  Fractional Weighted Bootstrap Covariance Matrix Estimation
+
+- [`fwb.ci()`](https://ngreifer.github.io/fwb/reference/fwb.ci.md)
+  [`print(`*`<fwbci>`*`)`](https://ngreifer.github.io/fwb/reference/fwb.ci.md)
+  : Fractional Weighted Bootstrap Confidence Intervals
+
 - [`plot(`*`<fwb>`*`)`](https://ngreifer.github.io/fwb/reference/plot.fwb.md)
   : Plots of the Output of a Fractional Weighted Bootstrap
 
-- [`vcovFWB()`](https://ngreifer.github.io/fwb/reference/vcovFWB.md) :
-  Fractional Weighted Bootstrap Covariance Matrix Estimation
+- [`get_ci()`](https://ngreifer.github.io/fwb/reference/get_ci.md) :
+
+  Extract Confidence Intervals from a `bootci` Object
+
+- [`set_fwb_wtype()`](https://ngreifer.github.io/fwb/reference/set_fwb_wtype.md)
+  [`get_fwb_wtype()`](https://ngreifer.github.io/fwb/reference/set_fwb_wtype.md)
+  : Set weights type
 
 - [`w_mean()`](https://ngreifer.github.io/fwb/reference/w_mean.md)
[…]
