### 2026-08-26 0cd5700
Bib updates for JOSS
