+++ joss_paper/paper.bib
@@ -3,7 +3,7 @@
   efronIntroductionBootstrap1993,
   place={New York},
   series={Monographs on statistics and applied probability},
-  title={An introduction to the bootstrap},
+  title={An Introduction to the Bootstrap},
   ISBN={978-0-412-04231-7},
   publisher={Chapman & Hall},
   author={Efron, Bradley and Tibshirani, Robert},
@@ -49,46 +49,44 @@
 	langid = {eng}
 }
 
-@misc{canty2024,
-	title = {boot: Bootstrap Functions (Originally by Angelo Canty for S)},
-	author = {Canty, Angelo and Ripley, Brian},
-	year = {2024},
-	publisher = {The R Foundation},
-	doi = {10.32614/cran.package.boot}
+@manual{canty2024,
+	title = {{boot}: {Bootstrap} Functions},
+    author = {Angelo Canty and Brian Ripley},
+    year = {2025},
+    note = {R package version 1.3-32},
+    url = {https://CRAN.R-project.org/package=boot},
+    doi = {10.32614/CRAN.package.boot},
 }
 
-@book{
-  davisonBootstrapMethodsTheir1997,
-  place={Cambridge},
-  edition={Reprinted with corrections 2003},
-  series={Cambridge series on statistical and probabilistic mathematics},
-  title={Bootstrap methods and their application},
-  ISBN={978-0-521-57391-7 978-0-521-57471-6},
-  publisher={Cambridge University Press},
-  author={Davison, Anthony Christopher and Hinkley, David Victor},
-  year={1997},
-  collection={Cambridge series on statistical and probabilistic mathematics}
+@book{davisonBootstrapMethodsTheir1997,
+  title = {Bootstrap Methods and Their Applications},
+    author = {Anthony Christopher Davison and David Victor Hinkley},
+    publisher = {Cambridge University Press},
+    address = {Cambridge},
+    year = {1997},
+    isbn = {0-521-57391-2},
+    doi = {10.1017/CBO9780511802843},
 }
 
-@misc{solymos2023,
-	title = {pbapply: Adding Progress Bar to '*apply' Functions},
-	author = {Solymos, Peter and Zawadzki, Zygmunt},
-	year = {2023},
-	publisher = {The R Foundation},
-	doi = {10.32614/cran.package.pbapply}
+@manual{solymos2023,
+	title = {{pbapply}: {Adding} Progress Bar to '*apply' Functions},
+    author = {Peter Solymos and Zygmunt Zawadzki},
+    year = {2025},
+    note = {R package version 1.7-4},
+    url = {https://CRAN.R-project.org/package=pbapply},
+    doi = {10.32614/CRAN.package.pbapply},
 }
 
 @article{bengtsson2021,
-	title = {A Unifying Framework for Parallel and Distributed Processing in R using Futures},
-	author = {Bengtsson, Henrik},
-	year = {2021},
-	date = {2021},
-	journal = {The R Journal},
-	pages = {208},
-	volume = {13},
-	number = {2},
-	doi = {10.32614/rj-2021-048},
-	langid = {en}
+	author = {Henrik Bengtsson},
+    title = {A Unifying Framework for Parallel and Distributed Processing in {R} using Futures},
+    year = {2021},
+    journal = {The R Journal},
+    doi = {10.32614/RJ-2021-048},
+    url = {https://doi.org/10.32614/RJ-2021-048},
+    pages = {208--227},
+    volume = {13},
+    number = {2},
 }
 
 @article{mammenBootstrapWildBootstrap1993,
@@ -131,27 +129,28 @@
 }
 
 @article{zeileis2020,
[…]
