+++ DESCRIPTION
@@ -46,4 +46,4 @@ LazyData: true
 Config/testthat/edition: 3
 Config/testthat/parallel: false
 VignetteBuilder: knitr
-Config/roxygen2/version: 8.0.0
+Config/roxygen2/version: 8.1.0
