### 2026-08-28 #8880 @crvernon
@ngreifer  - I just need you to address the following before I can accept this submission for publication:

In your archive:

- [ ] The title of your archive must match the title of your paper exactly.  Please edit the metadata of your archive to reflect this.  No new release required.
- [ ] I see you have listed the correct license but did not remove the default CC one. Please do this.

In your paper:

- [ ] LINE 47: The citation in “…Beta and Power weights described by (Owen, 2025)…” is incorrectly formatted. Instead “…Beta and Power weights described by Owen (2025)…”
- [ ] LINE 82: The “b” in “bayesian” should be capitalized because Bayesian derives from the proper name Bayes. Write “Bayesian bootstrap.”  You can do this by placing curly brackets around characters you wish to maintain the formatting of in your bib file.
- [ ] LINE 103: The “b” in “bayesian” should be capitalized

Let me know when these have been addressed.  Thank you.
