+++ joss_paper/paper.bib
@@ -12,7 +12,7 @@
 }
 
 @article{rubinBayesianBootstrap1981,
-	title = {The Bayesian Bootstrap},
+	title = {The {B}ayesian Bootstrap},
 	author = {Rubin, Donald B.},
 	year = {1981},
 	date = {1981},
@@ -140,7 +140,7 @@
 }
 
 @manual{bayesboot,
-    title = {{bayesboot}: {An} Implementation of {R}ubin's (1981) Bayesian Bootstrap},
+    title = {{bayesboot}: {An} Implementation of {R}ubin's (1981) {B}ayesian Bootstrap},
     author = {Rasmus Bååth},
     year = {2025},
     note = {R package version 0.2.3},
