+++ joss_paper/paper.md
@@ -25,7 +25,7 @@ The fractional weighted bootstrap was described originally by @rubinBayesianBoot
 
 This is demonstrated in @xuApplicationsFractionalRandomWeightBootstrap2020a, who use the fractional weighted bootstrap to estimate the center and shape of a Weibull distribution characterizing failure times for a set of 1703 aircraft engines, among which only 6 failures were observed, data previously analyzed by @meekerStatisticalMethodsReliability1998. In cases with rare outcomes, rare covariate profiles, or highly influential observations, the traditional bootstrap can fail to yield valid inference. The fractional weighted bootstrap avoids these issues by retaining all units and allowing the weights to take on strictly positive, non-integer values.
 
-The traditional bootstrap is implemented in the `boot` package [@canty2024; @davisonBootstrapMethodsTheir1997], which uses a simple interface to perform bootstrapping. The user supplies a dataset and a function that computes the desired statistics on each bootstrap sample from the dataset to `boot::boot()`. `boot::boot.ci()` is used to compute confidence intervals using a variety of methods. `fwb` provides a drop-in replacement for `boot::boot()` in `fwb::fwb()`, which similarly takes in the original dataset and a function that computes the desired statistics incorporating the bootstrap weights. `fwb()` supports parallelization through the `pbapply` package [@solymos2023], which provides interfaces to the `parallel` and `future` packages [@bengtsson2021] for parallel processing, as well as stratified bootstrapping (i.e., bootstrapping within strata), clustered bootstrapping (bootstrapping of clusters), and different distributions from which the weights are drawn. These distributions include the Dirichlet/Exponential weights originally described by @rubinBayesianBootstrap1981, "Mammen" weights that mimic the form of the distribution of weights proposed by @mammenBootstrapWildBootstrap1993 for the wild bootstrap, Beta and Power weights described by [@owenBetterBootstrapConfidence2025], multinomial weights to exactly reproduce the traditional nonparametric bootstrap by `boot`, and Poisson frequency weights [@hanley2006].
+The traditional bootstrap is implemented in the `boot` package [@canty2024; @davisonBootstrapMethodsTheir1997], which uses a simple interface to perform bootstrapping. The user supplies a dataset and a function that computes the desired statistics on each bootstrap sample from the dataset to `boot::boot()`. `boot::boot.ci()` is used to compute confidence intervals using a variety of methods. `fwb` provides a drop-in replacement for `boot::boot()` in `fwb::fwb()`, which similarly takes in the original dataset and a function that computes the desired statistics incorporating the bootstrap weights. `fwb()` supports parallelization through the `pbapply` package [@solymos2023], which provides interfaces to the `parallel` and `future` packages [@bengtsson2021] for parallel processing, as well as stratified bootstrapping (i.e., bootstrapping within strata), clustered bootstrapping (bootstrapping of clusters), and different distributions from which the weights are drawn. These distributions include the Dirichlet/Exponential weights originally described by @rubinBayesianBootstrap1981, "Mammen" weights that mimic the form of the distribution of weights proposed by @mammenBootstrapWildBootstrap1993 for the wild bootstrap, Beta and Power weights described by @owenBetterBootstrapConfidence2025, multinomial weights to exactly reproduce the traditional nonparametric bootstrap by `boot`, and Poisson frequency weights [@hanley2006].
 
[…]
